[ADD]: Stage parent class
