+++ src/phaser/scenes/singleplay/Preloader.js
@@ -49,7 +49,7 @@ export default class Preloader extends Phaser.Scene {
   }
 
   completeLoading() {
-    this.scene.start("stage1");
+    this.scene.start("stage3");
   }
   
   loadBackgroundResource() {
